added types.ts for common types
